producers/requirements.txt
grpcio==1.71.0
kafka-python==2.0.2
langdetect==1.0.9
orjson==3.9.15
protobuf>=4.25.1

